RSE-1388: Fix naming conventions
