+++ CRM/CiviAwards/Upgrader/Steps/Step1007.php
@@ -1,6 +1,6 @@
 <?php
 
-use CRM_CiviAwards_Setup_RenameAwardTypeField as RenameAwardTypeField;
+use CRM_CiviAwards_Setup_CreateAwardSubtypeOptionGroup as CreateAwardSubtypeOptionGroup;
 
 /**
  * Renames `civiawards_award_type` option group to `civiawards_award_subtype`.
@@ -14,10 +14,44 @@ class CRM_CiviAwards_Upgrader_Steps_Step1007 {
    *   returns value.
    */
   public function apply() {
-    $step = new RenameAwardTypeField();
-    $step->apply();
+    $this->renameDbColumn();
+    $this->renameOptionGroup();
 
     return TRUE;
+  }
+
+  /**
+   * Renames `civiawards_award_type` option group to `civiawards_award_subtype`.
+   */
+  private function renameOptionGroup() {
+    $awardSubtypeOptionGroup = new CreateAwardSubtypeOptionGroup();
+
+    $awardTypeOptionGroup = civicrm_api3('OptionGroup', 'get', [
+      'sequential' => 1,
+      'name' => "civiawards_award_type",
+    ])['values'];
+
+    if (count($awardTypeOptionGroup) == 0) {
+      return;
+    }
+
+    civicrm_api3('OptionGroup', 'create', [
+      'id' => $awardTypeOptionGroup[0]['id'],
+      'name' => $awardSubtypeOptionGroup->awardOptionGroupName,
+      'title' => ts($awardSubtypeOptionGroup->awardOptionGroupTitle),
+      'is_reserved' => 1,
+    ]);
+  }
+
+  /**
+   * Renames `award_type` column to `award_subtype`.
+   */
+  private function renameDbColumn() {
+    $renameColumnSql = "ALTER TABLE
+      civicrm_civiawards_award_detail
+      CHANGE award_type award_subtype varchar(30) NOT NULL COMMENT 'One of the values of the award_subtype option group'";
+
+    CRM_Core_DAO::executeQuery($renameColumnSql);
   }
 
 }
